requirements.txt
mysqlclient==1.4
django==2.1.5
djangorestframework==3.7.7